`concealwallet` command improvements
Over the years of development to client wallet, there have been many commands implemented. When a user types "help", there could be a lot to make a new user scared because of the shear amount of commands there is. I've split all the commands into two menu's, "help" and "ext_help".

The command "save_keys" has now been improved to be able to save multiple backups without overwriting the previous. In example: If I've made multiple wallets in concealwallet and decided to back each one up, we would end up with only one set of keys being back as each backup would overwrite the last. Now this function has been changed so it includes the wallet filename within the filename of the backed up keys. This also covers 1 SonarCloud code smell by making use of "args" in this command