README: Updated Dependencies & Other Stuff
--> Removed redundant dependencies
'gcc' and 'g++' are included in the 'build-essential' package, so also
listing them to be installed is redundant.

--> Changed 'python-dev' to 'python3-dev'
On Ubuntu and Debian Based Systems, 'python' refers to Python 2, and the
package that is actually needed is named 'python3-dev'

--> Changed 'Ubuntu' to 'Linux'
Changed the header of 'Building/Ubuntu' to 'Building/Linux', as the
commands there are pretty much identical across all distributions.

--> 'apt-get' is now 'apt'
In the dependencies, removed '-get' from 'apt-get install', as it's not
required, 'apt' works, and most people are just going to use plain 'apt'